fix: dependency restructuring
