+++ src/components/structural/index.ts
@@ -1,5 +1,4 @@
 export * from "./Header";
-export * from "./ButtonNav";
 export * from "./Footer";
 export * from "./FooterBottom";
 export * from "./Layout";
